+++ src/Collision.ts
@@ -2,57 +2,27 @@ import { CommonArea, CommonOffset } from "@akashic/pdi-types";
 import { E } from "./entities/E";
 import { Util } from "./Util";
 
-// TODO @akashic/collision-js を部分的に統合する？
-type Vec2 = CommonOffset;
-
-function dot(v1: Vec2, v2: Vec2): number {
-	return v1.x * v2.x + v1.y * v2.y;
+// 外積の絶対値
+function absCross(v1: CommonOffset, v2: CommonOffset): number {
+	return v1.x * v2.y - v1.y * v2.x;
 }
 
-function normalize(v: Vec2): Vec2 {
-	const len = Math.sqrt(dot(v, v));
-	v.x /= len;
-	v.y /= len;
-	return v;
+function sub(v1: CommonOffset, v2: CommonOffset): CommonOffset {
+	return { x: v1.x - v2.x, y: v1.y - v2.y };
 }
 
 /**
- * 線分と、線分に重なる直線に投影した矩形の交差判定。
- * @param lt 矩形の頂点の一つ(左上端)
- * @param rt 矩形の頂点の一つ(右上端)
- * @param lb 矩形の頂点の一つ(左下端)
- * @param rb 矩形の頂点の一つ(右下端)
- * @param p0 線分の一端
- * @param p1 線分の一端
- */
-function overlapBoxAndSegment(lt: Vec2, rt: Vec2, lb: Vec2, rb: Vec2, p0: Vec2, p1: Vec2): boolean {
-	const dir = normalize({ x: p1.x - p0.x, y: p1.y - p0.y });
-	const tp0 = dot(p0, dir);
-	const tp1 = dot(p1, dir);
-	const tlt = dot(lt, dir);
-	const trt = dot(rt, dir);
-	const tlb = dot(lb, dir);
-	const trb = dot(rb, dir);
-
-	const tpMin = Math.min(tp0, tp1);
-	const tpMax = Math.max(tp0, tp1);
-	const tbMin = Math.min(tlt, trt, tlb, trb);
-	const tbMax = Math.max(tlt, trt, tlb, trb);
-	return !(tbMax < tpMin || tpMax < tbMin);
-}
-
-/**
- * オブジェクトの衝突を表す。
- * - 矩形交差による衝突
- * - 2点間の距離による衝突
+ * オブジェクトの衝突判定機能を提供する。
+ *
+ * 距離による判定・矩形交差による判定、エンティティの親子関係を踏まえた低速だが厳密な判定
  */
 export module Collision {
 	/**
 	 * 二つのエンティティの衝突判定を行い、その結果を返す。
 	 *
 	 * 回転・拡大されたエンティティや、親の異なるエンティティ同士も扱える汎用の衝突判定処理。
-	 * ただし計算量が多いので、大量のオブジェクトとの判定には不向きである。
-	 * 親が同じで回転や拡大を行わないエンティティ同士の場合は、より軽量な Collision.intersectAreas() を利用すること。
+	 * ただし計算量が多いので、大量のエンティティ間のすべての衝突を確認するような状況では利用を避けることが望ましい。
+	 * 親が同じで回転・拡大を行わないエンティティ同士の場合は、より軽量な Collision.intersectAreas() を利用すること。
 	 * 親が同じで中心座標同士の距離だけで判定してよい場合は、より軽量な Collision.withinAreas() を利用すること。
 	 *
 	 * 対象のエンティティの座標や大きさなどを変更した場合、
@@ -67,9 +37,6 @@ export module Collision {
 		const lca = e1.findLowestCommonAncestorWith(e2);
 		if (!lca) return false;
 
-		const mat1 = e1.calculateMatrixTo(lca);
-		const mat2 = e2.calculateMatrixTo(lca);
-
 		const r1 = area1
 			? { left: area1.x, top: area1.y, right: area1.x + area1.width, bottom: area1.y + area1.height }
 			: { left: 0, top: 0, right: e1.width, bottom: e1.height };
@@ -77,7 +44,10 @@ export module Collision {
 			? { left: area2.x, top: area2.y, right: area2.x + area2.width, bottom: area2.y + area2.height }
 			: { left: 0, top: 0, right: e2.width, bottom: e2.height };
 
-		// LCA の座標系に合わせたそれぞれの四隅の点
+		const mat1 = e1.calculateMatrixTo(lca);
+		const mat2 = e2.calculateMatrixTo(lca);
+
+		// 座標系を合わせる: 共通祖先の座標系に合わせたそれぞれの四隅の点を求める。
 		const lt1 = mat1.multiplyPoint({ x: r1.left, y: r1.top });
 		const rt1 = mat1.multiplyPoint({ x: r1.right, y: r1.top });
 		const lb1 = mat1.multiplyPoint({ x: r1.left, y: r1.bottom });
@@ -87,25 +57,90 @@ export module Collision {
 		const lb2 = mat2.multiplyPoint({ x: r2.left, y: r2.bottom });
 		const rb2 = mat2.multiplyPoint({ x: r2.right, y: r2.bottom });
 
+		// AABB で枝狩りする。(高速化だけでなく後続の条件を単純化するのにも必要である点に注意)
+		const minX1 = Math.min(lt1.x, rt1.x, lb1.x, rb1.x);
+		const maxX1 = Math.max(lt1.x, rt1.x, lb1.x, rb1.x);
+		const minX2 = Math.min(lt2.x, rt2.x, lb2.x, rb2.x);
+		const maxX2 = Math.max(lt2.x, rt2.x, lb2.x, rb2.x);
+		if (maxX1 < minX2 || maxX2 < minX1) return false;
+		const minY1 = Math.min(lt1.y, rt1.y, lb1.y, rb1.y);
+		const maxY1 = Math.max(lt1.y, rt1.y, lb1.y, rb1.y);
+		const minY2 = Math.min(lt2.y, rt2.y, lb2.y, rb2.y);
+		const maxY2 = Math.max(lt2.y, rt2.y, lb2.y, rb2.y);
+		if (maxY1 < minY2 || maxY2 < minY1) return false;
+
+		// 二つの四角形それぞれのいずれかの辺同士が交差するなら衝突している。
+		if (
+			Collision.intersectLineSegments(lt1, rt1, lt2, rt2) ||
+			Collision.intersectLineSegments(lt1, rt1, rt2, rb2) ||
+			Collision.intersectLineSegments(lt1, rt1, rb2, lb2) ||
+			Collision.intersectLineSegments(lt1, rt1, lb2, lt2) ||
+			Collision.intersectLineSegments(rt1, rb1, lt2, rt2) ||
+			Collision.intersectLineSegments(rt1, rb1, rt2, rb2) ||
+			Collision.intersectLineSegments(rt1, rb1, rb2, lb2) ||
+			Collision.intersectLineSegments(rt1, rb1, lb2, lt2) ||
+			Collision.intersectLineSegments(rb1, lb1, lt2, rt2) ||
+			Collision.intersectLineSegments(rb1, lb1, rt2, rb2) ||
+			Collision.intersectLineSegments(rb1, lb1, rb2, lb2) ||
+			Collision.intersectLineSegments(rb1, lb1, lb2, lt2) ||
+			Collision.intersectLineSegments(lb1, lt1, lt2, rt2) ||
+			Collision.intersectLineSegments(lb1, lt1, rt2, rb2) ||
+			Collision.intersectLineSegments(lb1, lt1, rb2, lb2) ||
+			Collision.intersectLineSegments(lb1, lt1, lb2, lt2)
+		) {
+			return true;
+		}
+
+		// そうでない場合、e1 が e2 を包含しているなら衝突している。
+		// ここで辺は交差していないので、e1 が e2 の頂点一つ (lt2) を包含しているなら、全体を包含している。
+		// cf. https://ksta.skr.jp/topic/diaryb09.html#040528 "各辺の内側判定による内外判定"
+		const s1 = absCross(sub(lt1, rt1), sub(lt2, rt1));
+		if (
+			s1 * absCross(sub(lb1, lt1), sub(lt2, lt1)) >= 0 &&
+			s1 * absCross(sub(rb1, lb1), sub(lt2, lb1)) >= 0 &&
+			s1 * absCross(sub(rt1, rb1), sub(lt2, rb1)) >= 0
+		) {
+			return true;
+		}
+
+		// そうでない場合、e2 が e1 を包含しているなら衝突している。
+		const s2 = absCross(sub(lt2, rt2), sub(lt1, rt2));
 		return (
-			overlapBoxAndSegment(lt1, rt1, lb1, rb1, lt2, rt2) &&
-			overlapBoxAndSegment(lt1, rt1, lb1, rb1, lt2, lb2) &&
-			overlapBoxAndSegment(lt2, rt2, lb2, rb2, lt1, rt1) &&
-			overlapBoxAndSegment(lt2, rt2, lb2, rb2, lt1, lb1)
+			s2 * absCross(sub(lb2, lt2), sub(lt1, lt2)) >= 0 &&
+			s2 * absCross(sub(rb2, lb2), sub(lt1, lb2)) >= 0 &&
+			s2 * absCross(sub(rt2, rb2), sub(lt1, rb2)) >= 0
+		);
+	}
+
+	/**
+	 * 線分同士の衝突判定 (交差判定) を行い、その結果を返す。
+	 *
+	 * @param {CommonOffset} p1 線分の端点の一つ
+	 * @param {CommonOffset} p2 線分の端点の一つ
+	 * @param {CommonOffset} q1 もう一つの線分の端点の一つ
+	 * @param {CommonOffset} q2 もう一つの線分の端点の一つ
+	 */
+	export function intersectLineSegments(p1: CommonOffset, p2: CommonOffset, q1: CommonOffset, q2: CommonOffset): boolean {
+		// cf. https://ksta.skr.jp/topic/diaryb09.html#040518
+		const p = sub(p2, p1);
+		const q = sub(q2, q1);
+		return (
+			absCross(sub(q1, p1), p) * absCross(sub(q2, p1), p) <= 0 && absCross(sub(p1, q1), q) * absCross(sub(p2, q1), q) <= 0 // 符号が違うことを積の符号で判定している
 		);
 	}
 
 	/**
 	 * 矩形交差による衝突判定を行い、その結果を返す。
-	 * 戻り値は、矩形t1, t2が交差しているとき真、でなければ偽。
-	 * @param {number} x1 t1-X
-	 * @param {number} y1 t1-Y
-	 * @param {number} width1 t1幅
-	 * @param {number} height1 t1高さ
-	 * @param {number} x2 t2-X
-	 * @param {number} y2 t2-Y
-	 * @param {number} width2 t2幅
-	 * @param {number} height2 t2高さ
+	 * 戻り値は、二つの矩形t1, t2が交差しているとき真、でなければ偽。
+	 *
+	 * @param {number} x1 t1のX座標
+	 * @param {number} y1 t1のY座標
+	 * @param {number} width1 t1の幅
+	 * @param {number} height1 t1の高さ
+	 * @param {number} x2 t2のX座標
+	 * @param {number} y2 t2のY座標
+	 * @param {number} width2 t2の幅
+	 * @param {number} height2 t2の高さ
 	 */
 	export function intersect(
 		x1: number,
@@ -123,6 +158,10 @@ export module Collision {
 	/**
 	 * 矩形交差による衝突判定を行い、その結果を返す。
 	 * 戻り値は、矩形t1, t2が交差しているとき真、でなければ偽。
+	 *
+	 * 特に、回転・拡大を利用していない、親が同じエンティティ同士の衝突判定に利用することができる。
+	 * 条件を満たさない場合は `withinAreas()` や、より重いが正確な `intersectEntities()` の利用を検討すること。
+	 *
 	 * @param {CommonArea} t1 矩形1
 	 * @param {CommonArea} t2 矩形2
 	 */